pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hexlet-code"
version = "0.1.0"
description = "Compares two configuration files and shows a difference."
readme = "README.md"
requires-python = ">=3.8"
authors = [
    {name = "Ilya Lysenkov", email = "lysenkov_iv21@mail.ru"},
]
dependencies = [
    "PyYAML>=6.0",
]

[tool.hatch.build.targets.wheel]
packages = ["gendiff"]

[project.scripts]
gendiff = "gendiff.scripts.gendiff:main"
lint = "flake8 gendiff tests"
test = "pytest -v --cov=gendiff"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0",
]